# web/src/api/openapi.json
{
  "components": {
    "schemas": {
      "AuthErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "AuthErrorResponse",
        "type": "object"
      },
...
      "ChallengeCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "target_channel": {
            "enum": [
              "telegram",
              "wechat"
            ],
            "title": "Target Channel",
            "type": "string"
          }
        },
        "required": [
          "target_channel"
        ],
        "title": "ChallengeCreateRequest",
        "type": "object"
      },
      "ChallengeCreateResponse": {
        "additionalProperties": false,
        "properties": {
          "browser_secret": {
            "title": "Browser Secret",
            "type": "string"
          },
          "command": {
            "title": "Command",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "target_channel": {
            "enum": [
              "telegram",
              "wechat"
            ],
            "title": "Target Channel",
            "type": "string"
          }
        },
        "required": [
          "public_id",
          "command",
          "browser_secret",
          "target_channel",
          "expires_at"
        ],
        "title": "ChallengeCreateResponse",
        "type": "object"
      },
      "ChallengeReferenceRequest": {
        "additionalProperties": false,
        "properties": {
          "public_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Public Id",
            "type": "string"
          }
        },
        "required": [
          "public_id"
        ],
        "title": "ChallengeReferenceRequest",
        "type": "object"
      },
      "ChallengeStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "approved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "expires_at"
        ],
        "title": "ChallengeStatusResponse",
        "type": "object"
      },
...
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "const": true,
            "default": true,
            "title": "Authenticated",
            "type": "boolean"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "login_channel": {
            "enum": [
              "telegram",
              "wechat"
            ],
            "title": "Login Channel",
            "type": "string"
          }
        },
        "required": [
          "login_channel",
          "expires_at"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BrowserSecret": {
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "in": "cookie",
        "name": "__Host-kb_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Notebook Agent Web API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/challenges": {
      "post": {
        "operationId": "create_challenge_api_v1_auth_challenges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Create Challenge",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/challenges/status": {
      "post": {
        "operationId": "challenge_status_api_v1_auth_challenges_status_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeReferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeStatusResponse"
                }
              }
            },
            "description": "Accepted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "BrowserSecret": []
          }
        ],
        "summary": "Challenge Status",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/session": {
      "delete": {
        "operationId": "logout_api_v1_auth_session_delete",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "authentication"
        ]
      },
      "get": {
        "operationId": "current_session_api_v1_auth_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Current Session",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "post": {
        "operationId": "exchange_session_api_v1_auth_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeReferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Accepted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "BrowserSecret": []
          }
        ],
        "summary": "Exchange Session",
        "tags": [
          "authentication"
        ]
      }
    },
...
  }
}